move ID's to helper class
